+++ src/main/java/com/wkclz/core/pojo/entity/AccessLog.java
@@ -163,12 +163,12 @@ public class AccessLog {
     /**
      * 认证
      */
-    private Integer authId;
+    private Long authId;
 
     /**
      * 用户ID
      */
-    private Integer userId;
+    private Long userId;
 
     /**
      * 用户昵称
